zist-backend/requirements.txt
fastapi[standard]==0.116.1
SQLAlchemy==2.0.42
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
email-validator==2.2.0
psycopg2-binary==2.9.10
psycopg[binary]==3.2.9
google-generativeai==0.7.2
